[Either] Remove optional extension and replace to map
