+++ core/Sources/Components/Switch/Model/Public/Images/SwitchImages+ExtensionTests.swift
@@ -1,0 +1,16 @@
+//
+//  SwitchImages+ExtensionTests.swift
+//  SparkCoreTests
+//
+//  Created by robin.lemaire on 03/07/2023.
+//  Copyright © 2023 Adevinta. All rights reserved.
+//
+
+@testable import SparkCore
+
+extension SwitchImages: Equatable {
+
+    public static func == (lhs: SwitchImages, rhs: SwitchImages) -> Bool {
+        return lhs.on == rhs.on && lhs.off == rhs.off
+    }
+}
