+++ core/Sources/Components/Switch/Model/Public/Images/SwitchImages.swift
