+++ core/Sources/Components/Switch/Model/Public/Images/SwitchUIImages.swift
